use path.posix when dealing with manta paths
this fixes a bug on windows when a dir is lazily created using
`client.put()`

ie. `client.put('/foo/bar/baz')`: If this yields a
`DirectoryDoesNotExistError`, it will attempt to:

```client.mkdirp(path.dirname(path.normalize('/foo/bar/baz')))```

(as expected).

However, on Windows, this will result in `client.mkdirp('\\foo\\bar')`,
which will fail when the request is made to manta.  This issue can be
avoided by using `path.posix` for any and all path related calls that
deal with manta paths.